picker/requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
pandas
pytz
pyodbc
requests
gunicorn
